requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
ortools==9.8.3296
networkx==3.2.1
pydantic==2.6.1
pandas==2.2.0
plotly==5.18.0
python-dotenv==1.0.1
holidays==0.40
PyPDF2==3.0.1
python-docx==1.1.0